==> 1_Maturity.ipynb <==
# Goal of the notebook

In order to apply classical models (such as Black-Scholes model which we will use as a benchmark), we need to create some very basic features (such as time until expiration and risk-free rate). In this notebook, we will do it.

import numpy as np
import pandas as pd

dataset_full = pd.read_csv('Raw_data_02_05.csv')
dataset_full

### Subtract expiration date from date of collection

dataset_full[['Exp_date', 'Exp_year']] = dataset_full['Expiration date'].str.split(',', expand=True)
dataset_full[['Exp_month', 'Exp_day']] = dataset_full['Exp_date'].str.split(' ', expand=True)
diction = {"January" : 1, "February" : 2, "March": 3, "April" : 4, 
           "May": 5, 'June': 6, 'July': 7, 'August': 8,
          'September': 9, 'October': 10, 'November': 11, 'December': 12}
dataset_full = dataset_full.replace({"Exp_month": diction})
dataset_full['Get_year'] = 2024
dataset_full[['Get_day', 'Get_month']] = dataset_full['Date of getting data'].str.split(' ', expand=True)
dataset_full = dataset_full.replace({"Get_month": diction})
dataset_full

dataset_full['Date_get_data'] = pd.to_datetime(dict(year=dataset_full.Get_year, month=dataset_full.Get_month, day=dataset_full.Get_day))
dataset_full['Date_expiration'] = pd.to_datetime(dict(year=dataset_full.Exp_year, month=dataset_full.Exp_month, day=dataset_full.Exp_day))
dataset_full = dataset_full.drop(['Exp_day', 'Exp_year', 'Exp_month', 'Get_year', 'Get_day', 'Get_month',
                                 'Date of getting data', 'Expiration date',
                                 'Last Trade Date', 'Exp_date'], axis=1)

dataset_full['Time_to_maturity'] = (dataset_full['Date_expiration'] - dataset_full['Date_get_data']).dt.days
dataset_full['Time_to_maturity'] = dataset_full['Time_to_maturity'] / 365 # do as in our source
dataset_full['r'] = 0.04 # assume it to be fixed and to be 4%
dataset_full.shape

dataset_full = dataset_full.loc[dataset_full['Contract Name'].notna()]
dataset_full.shape

dataset_full.head()

dataset_full.to_csv('Maturity_data_02_05.csv', index=None)

### Complete
