==> 2_Merge_with_stock_data.ipynb <==
# Goal of the notebook

So far we have collected only data related to options themselves. However, in order to apply Black-Scholes model (and probably many other reasonable models), we need to have price of an underlying assets as inputs. Hence, the goal of this notebook is to merge our option data with stock data. We will also extract trading volumes of underlying assets as additional features.

import pandas as pd
import numpy as np

dataset_full = pd.read_csv('Maturity_data_02_05.csv')
dataset_full['Date_get_data'] = pd.to_datetime(dataset_full['Date_get_data'])
dataset_full['Date_expiration'] = pd.to_datetime(dataset_full['Date_expiration'])
dataset_full

dataset_full.dtypes

### Get prices and trading volumes

import yfinance as yf
import datetime as dt
import pandas_datareader as pdr
pd.set_option('display.max_columns', None)

yf.pdr_override()

from pandas_datareader import data as pdr 
import matplotlib.pyplot as plt

start = "2020-01-01"
end = "2024-04-13"

symbols_stocks = ['AAPL', 
            'GOOGL',
            'MSFT',
            'SPY',
            'PFE',
            'LMT',
            'DIS',
            'VZ',
            'PYPL',
            'TSN',
            'DG',
            'QCOM',
            'META',
            'TSLA',
            'NFLX',
            'NVDA',
            'AMZN',
            'AVGO',
            'LLY',
            'JPM',
            'UNH',
            'V',
            'MA',
            'JNJ',
            'PG',
            'MRK',
            'MCD',
            'WMT',
            'BAC',
            'ACN',
            'KO',
            'PEP',
            'CSCO',
            'ORCL',
            'INTC'] 

data_stocks_close = pdr.get_data_yahoo(symbols_stocks, start, end)['Close']
data_stocks_volume = pdr.get_data_yahoo(symbols_stocks, start, end)['Volume']

# prices of underlying assets
data_stocks_close = data_stocks_close.reset_index()
data_stocks_close['Date'] = pd.to_datetime(data_stocks_close['Date'])
data_stocks_close

# trading volumes
data_stocks_volume = data_stocks_volume.reset_index()
data_stocks_volume['Date'] = pd.to_datetime(data_stocks_volume['Date'])
data_stocks_volume

### Merge with close price

res_merging = pd.DataFrame()
for symbol in symbols_stocks:
    for_merging = dataset_full.copy()
    for_merging = for_merging[['Date_get_data', 'Company']]
    for_merging = for_merging.drop_duplicates()
    tmp = pd.DataFrame(data_stocks_close[['Date', symbol]])
    tmp.columns = ['Date_get_data', 'Stock_price']
    tmp['Company'] = symbol
    for_merging = for_merging.merge(tmp, on=['Date_get_data', 'Company'], how='inner')
    res_merging = res_merging.append(for_merging)

dataset_full = dataset_full.merge(res_merging, on=['Date_get_data', 'Company'], how='left')

dataset_full.isnull().sum()

### Why nan?

Note that there is a column named Date_get_data. Sometimes I was getting data during weekends (meaning that I collected data on the previous working day compared to the date of getting data). Hence, the data was not available exactly on the data of collecting data, and we can just use ffill to fillna

set(dataset_full.loc[dataset_full['Stock_price'].isnull()]['Date_get_data'].values)

# Fillna smartly
dataset_full['Stock_price'] = dataset_full.groupby('Company')['Stock_price'].apply(lambda x: x.ffill())

dataset_full.isnull().sum()

### Merge with volume

res_merging = pd.DataFrame()
for symbol in symbols_stocks:
    for_merging = dataset_full.copy()
    for_merging = for_merging[['Date_get_data', 'Company']]
    for_merging = for_merging.drop_duplicates()
    tmp = pd.DataFrame(data_stocks_volume[['Date', symbol]])
    tmp.columns = ['Date_get_data', 'Stock_volume']
    tmp['Company'] = symbol
    for_merging = for_merging.merge(tmp, on=['Date_get_data', 'Company'], how='inner')
    res_merging = res_merging.append(for_merging)

dataset_full = dataset_full.merge(res_merging, on=['Date_get_data', 'Company'], how='left')

dataset_full.isnull().sum()

# Fillna smartly
dataset_full['Stock_volume'] = dataset_full.groupby('Company')['Stock_volume'].apply(lambda x: x.ffill())

dataset_full.isnull().sum()

dataset_full.head()

# checked, units are OK
dataset_full['Implied Volatility'] = dataset_full['Implied Volatility'].str.replace('%', '')
dataset_full['Implied Volatility'] = dataset_full['Implied Volatility'].str.replace(',', '')
dataset_full['Implied Volatility'] = dataset_full['Implied Volatility'].astype(float)
dataset_full['Implied Volatility'] = dataset_full['Implied Volatility'] * 0.01
dataset_full

# It is sometimes possible that I got data the day after the option has expired
dataset_full.loc[dataset_full['Time_to_maturity'] < 0, 'Time_to_maturity'] = 0

dataset_full.to_csv('3_Clean_data_02_05.csv', index=None)

### Create our own volatility

It is not fair to use implied volatility, since it is estimated with the use of models (such as Black-Scholes). Hence, using implied volatility means using data leakage (and we will see that later). Since implied volatility will not be available on normal test data, we will create our own volatility by taking variance of stock dynamics of underlying assets

volatilites = pd.DataFrame(data_stocks_close.describe())
volatilites

# Note that we do not square standard deviation
# This is done according to the Black-Scholes paper
volatilites = volatilites.loc[volatilites.index == 'std']
volatilites = volatilites.T
volatilites = volatilites.reset_index()
volatilites.columns = ['Company', 'Volatility']
volatilites

# however, later on we will not use it 
# (we will calculate volatilities based on contract themselves, not the underlying stocks)
volatilites.to_csv('volatilities_companies.csv', index=None)

1 + 1

### Complete
