==> 3_Filter_DATA.ipynb <==
# Goal of the notebook

We need to filter the data to make the model perform better. Initially I planned to collect slightly more than 1 million observations, so that we will get approximately 1 million observations in the end. I guess this will be sufficient to construct a fairly good model.

import pandas as pd
import numpy as np

dataset_full = pd.read_csv('3_Clean_data_02_05.csv')
dataset_full

def filter_contract_names(k, df=dataset_full):
    
    '''
    Leave only contract that we find at least k times
    '''
    contract_counts = df['Contract Name'].value_counts()
    valid_contract_names = contract_counts[contract_counts >= k].index
    filtered_df = df[df['Contract Name'].isin(valid_contract_names)]
    filtered_df.index = np.arange(0, len(filtered_df))
    return filtered_df

dataset_full_2 = filter_contract_names(11)
dataset_full_2

# If we leave only options that are represented at least 11 times in the dataset,
# We will lose less than 60 000 observations, which is equivalent to two days of collecting data
# However, we are now able to introduce 10 lag features for each option
len(dataset_full) - len(dataset_full_2)

dataset_full_2.to_csv('4_Filtered_data_02_05.csv', index=None)

### Complete
