==> 5_Benchmark_BS_LinReg.ipynb <==
# Goal of the notebook

Here we will apply Black-Scholes formula and Linear Regression as our benchmark and our baseline, respectively. Moreover, we will see that it is unfair to use implied volatility as a feature, as it is calculated either by using Black-Scholes model or a model very similar to it. As a result, we will stick with our own version of volatility.

We will fix the results obtained in this section. Therefore, we will be able to figure out, whether additional features and more advanced models will be helpful in the future.

import pandas as pd
import numpy as np
from scipy.stats import norm

# https://quantpy.com.au/black-scholes-model/implementing-black-scholes-option-pricing/
def blackScholes(r, S, K, T, sigma, type="c"):
    "Calculate BS price of call/put"
    d1 = (np.log(S/K) + (r + sigma**2/2)*T)/(sigma*np.sqrt(T))
    d2 = d1 - sigma*np.sqrt(T)
    if type == "c":
        price = S*norm.cdf(d1, 0, 1) - K*np.exp(-r*T)*norm.cdf(d2, 0, 1)
    elif type == "p":
        price = K*np.exp(-r*T)*norm.cdf(-d2, 0, 1) - S*norm.cdf(-d1, 0, 1)
    return price

dataset_full = pd.read_csv('4_Filtered_data_02_05.csv')
dataset_full

%%time
dataset_full['Black_Scholes'] = 0
for i in range(len(dataset_full)):
    dataset_full['Black_Scholes'][i] = blackScholes(r=0.04, S=dataset_full['Stock_price'][i],
                                                   K=dataset_full['Strike'][i], 
                                                    T=dataset_full['Time_to_maturity'][i],
                                                   sigma=dataset_full['Implied Volatility'][i])

dataset_full.isna().sum()

from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import mean_absolute_percentage_error

MSE_Black_Scholes = mean_squared_error(dataset_full['Last Price'], dataset_full['Black_Scholes'])
MAE_Black_Scholes = mean_absolute_error(dataset_full['Last Price'], dataset_full['Black_Scholes'])
MAPE_Black_Scholes = mean_absolute_percentage_error(dataset_full['Last Price'], dataset_full['Black_Scholes'])

# We will see later that this quality is very good compared to Linear Regression 
# and Black-Scholes on another volatility

MSE_Black_Scholes, MAE_Black_Scholes, MAPE_Black_Scholes

dataset_full.head()

### Linear Regression (only features used in Black-Scholes)

X = dataset_full[['Stock_price', 'Strike', 'Time_to_maturity', 'Implied Volatility']]
y = dataset_full[['Last Price']]

import statsmodels.api as sm

%%time
X_constant = sm.add_constant(X)
lin_reg = sm.OLS(y, X_constant).fit()
lin_reg.summary()

from sklearn.linear_model import LinearRegression

%%time
model = LinearRegression()
model.fit(X, y)
y_pred = model.predict(X)

MSE_linreg_resticted = mean_squared_error(y, y_pred)
MAE_linreg_resticted = mean_absolute_error(y, y_pred)
MAPE_linreg_resticted = mean_absolute_percentage_error(y, y_pred)

MSE_Black_Scholes, MAE_Black_Scholes, MAPE_Black_Scholes

# We see that quality of this model is dramatically worse
MSE_linreg_resticted, MAE_linreg_resticted, MAPE_linreg_resticted

### Create our own volatility instead of implied (using stock data)

volatilities_contract = pd.DataFrame(dataset_full.groupby('Contract Name')['Last Price'].std())
volatilities_contract = volatilities_contract.reset_index()
volatilities_contract.columns = ['Contract Name', 'Option_price_volatility']
volatilities_contract

dataset_full['Contract Name'].nunique()

dataset_full_1 = dataset_full.copy()
dataset_full_1 = dataset_full_1.drop('Implied Volatility', axis=1)
dataset_full_1

dataset_full_1 = dataset_full_1.merge(volatilities_contract, on='Contract Name', how='inner')
dataset_full_1.isnull().sum()

dataset_full_1['Black_Scholes'] = 0
for i in range(len(dataset_full_1)):
    dataset_full_1['Black_Scholes'][i] = blackScholes(r=0.04, S=dataset_full_1['Stock_price'][i],
                                                   K=dataset_full_1['Strike'][i], 
                                                    T=dataset_full_1['Time_to_maturity'][i],
                                                   sigma=dataset_full_1['Option_price_volatility'][i])

MSE_Black_Scholes_1 = mean_squared_error(dataset_full_1['Last Price'], dataset_full_1['Black_Scholes'])
MAE_Black_Scholes_1 = mean_absolute_error(dataset_full_1['Last Price'], dataset_full_1['Black_Scholes'])
MAPE_Black_Scholes_1 = mean_absolute_percentage_error(dataset_full_1['Last Price'], dataset_full_1['Black_Scholes'])


MSE_Black_Scholes, MAE_Black_Scholes, MAPE_Black_Scholes

MSE_Black_Scholes_1, MAE_Black_Scholes_1, MAPE_Black_Scholes_1

X_1 = dataset_full_1[['Stock_price', 'Strike', 'Time_to_maturity', 'Option_price_volatility']]
y_1 = dataset_full_1[['Last Price']]

model1 = LinearRegression()
model1.fit(X_1, y_1)
y_pred_1 = model.predict(X_1)

MSE_linreg_resticted_1 = mean_squared_error(y_1, y_pred_1)
MAE_linreg_resticted_1 = mean_absolute_error(y_1, y_pred_1)
MAPE_linreg_resticted_1 = mean_absolute_percentage_error(y_1, y_pred_1)

MSE_linreg_resticted, MAE_linreg_resticted, MAPE_linreg_resticted

MSE_linreg_resticted_1, MAE_linreg_resticted_1, MAPE_linreg_resticted_1

# Check time required for inference (predictions on new data)

X_test = X_1.sample(frac=.1)
X_test.index = np.arange(0, len(X_test))
X_test

%%time
# Time for Linear Regression
model1.predict(X_test)

%%time
# Time for Black-Scholes
X_test['Black_Scholes'] = 0
for i in range(len(X_test)):
    X_test['Black_Scholes'][i] = blackScholes(r=0.04, S=X_test['Stock_price'][i],
                                                   K=X_test['Strike'][i], 
                                                    T=X_test['Time_to_maturity'][i],
                                                   sigma=X_test['Option_price_volatility'][i])

# Conclusions

+ Implied volatility is a data leakage for Black-Scholes formula, since it is calculated either using Black-Scholes or a quite similar model.


+ Black-Scholes model is less consistent in terms of changing the way volatility is calculated.


+ Black-Scholes formula beats Linear Regression on the same set of features.


+ However, we need additional features to test the power of Linear Regression (and other Machine Learning methods).

dataset_full_1.to_csv('5_Fair_volatility_data.csv', index=None)
