==> 6_Feature_generation.ipynb <==
# Goal of the notebook

This notebook is used to create more advanced features. For now we are sure that additional features will improve the performance of models (especially time series features, such as lags). Besides them, we will add features related to companies (such as their age or weight in index). As a result, we will end up with a new version of the dataset.

import pandas as pd
import numpy as np

dataset_full = pd.read_csv('5_Fair_volatility_data.csv')
dataset_full

dataset_full['Contract Name'].nunique()

dataset_full.groupby('Contract Name')

### Get features related to companies

https://www.slickcharts.com/sp500 - weights

https://companiesmarketcap.com - capitalizations

Ages - information might not be very precise (but still the best we can do)

Another limitation of study - some useful features (news, dynamics of stocks in nearest past) are not used

# capitalizations in $billions
market_capitalizations = [['AAPL', 2831.0], 
            ['GOOGL', 2076.0],
            ['MSFT', 3022.0],
            ['SPY', 0],
            ['PFE', 157.47],
            ['LMT', 110.82],
            ['DIS', 208.48],
            ['VZ', 163.69],
            ['PYPL', 68.72],
            ['TSN', 22.14],
            ['DG', 30.20],
            ['QCOM', 200.47],
            ['META', 1146.0],
            ['TSLA', 577.85],
            ['NFLX', 249.63],
            ['NVDA', 2219.0],
            ['AMZN', 1937.0],
            ['AVGO', 592.30],
            ['LLY', 698.51],
            ['JPM', 547.08],
            ['UNH', 453.09],
            ['V', 549.35],
            ['MA', 412.35],
            ['JNJ', 359.24],
            ['PG', 388.14],
            ['MRK', 322.98],
            ['MCD', 194.90],
            ['WMT', 482.13],
            ['BAC', 291.30],
            ['ACN', 190.95],
            ['KO', 267.82],
            ['PEP', 242.16],
            ['CSCO', 190.79],
            ['ORCL', 318.27],
            ['INTC', 133.71]]

capitalizations = pd.DataFrame(market_capitalizations, columns=['Company', 'Capitalization'])
capitalizations.head()

# substitute 0 for S&P 500 with the median
med_capitalization = capitalizations['Capitalization'].median()
capitalizations.loc[capitalizations['Company'] == 'SPY', 'Capitalization'] = med_capitalization
capitalizations.head()

ages_company = [['AAPL', 48], 
            ['GOOGL', 26],
            ['MSFT', 49],
            ['SPY', 0],
            ['PFE', 175],
            ['LMT', 29],
            ['DIS', 101],
            ['VZ', 24],
            ['PYPL', 26],
            ['TSN', 31],
            ['DG', 85],
            ['QCOM', 39],
            ['META', 20],
            ['TSLA', 21],
            ['NFLX', 27],
            ['NVDA', 31],
            ['AMZN', 29],
            ['AVGO', 63],
            ['LLY', 148],
            ['JPM', 24],
            ['UNH', 57],
            ['V', 65],
            ['MA', 64],
            ['JNJ', 138],
            ['PG', 193],
            ['MRK', 133],
            ['MCD', 84],
            ['WMT', 62],
            ['BAC', 26],
            ['ACN', 35],
            ['KO', 137],
            ['PEP', 131],
            ['CSCO', 40],
            ['ORCL', 47],
            ['INTC', 56]]

ages = pd.DataFrame(ages_company, columns=['Company', 'Age'])
ages.head()

# substitute 0 for S&P 500 with the median 
med_ages = ages['Age'].median()
ages.loc[ages['Company'] == 'SPY', 'Age'] = med_ages
ages.head()

# weights in S&P 500 in %
weights_in_index = [['AAPL', 5.91], 
            ['GOOGL', 2.31],
            ['MSFT', 6.95],
            ['SPY', 0],
            ['PFE', 0.37],
            ['LMT', 0.23],
            ['DIS', 0.48],
            ['VZ', 0.38],
            ['PYPL', 0.17],
            ['TSN', 0.04],
            ['DG', 0.07],
            ['QCOM', 0.47],
            ['META', 2.29],
            ['TSLA', 1.17],
            ['NFLX', 0.58],
            ['NVDA', 4.98],
            ['AMZN', 3.97],
            ['AVGO', 1.28],
            ['LLY', 1.42],
            ['JPM', 1.30],
            ['UNH', 1.07],
            ['V', 1.00],
            ['MA', 0.86],
            ['JNJ', 0.85],
            ['PG', 0.91],
            ['MRK', 0.76],
            ['MCD', 0.47],
            ['WMT', 0.60],
            ['BAC', 0.60],
            ['ACN', 0.44],
            ['KO', 0.57],
            ['PEP', 0.57],
            ['CSCO', 0.45],
            ['ORCL', 0.43],
            ['INTC', 0.30]]

weights = pd.DataFrame(weights_in_index, columns=['Company', 'Weight'])
weights.head()

# substitute 0 for S&P 500 with the median
med_weight = weights['Weight'].median()
weights.loc[weights['Company'] == 'SPY', 'Weight'] = med_weight
weights.head()

add_features = capitalizations.merge(ages, on='Company')
add_features = add_features.merge(weights, on='Company')

# check that there were no mistakes while merging
assert len(add_features) == len(weights) == len(capitalizations) == len(ages), 'MISMATCH'

add_features.head()

dataset_full.head()

length = len(dataset_full)
dataset_full = dataset_full.merge(add_features, on='Company')
assert len(dataset_full) == length, 'MISMATCH'

### Create lag features and fillna smartly

dataset_full['l1_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(1)
dataset_full['l2_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(2)
dataset_full['l3_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(3)
dataset_full['l4_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(4)
dataset_full['l5_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(5)
dataset_full['l6_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(6)
dataset_full['l7_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(7)
dataset_full['l8_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(8)
dataset_full['l9_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(9)
dataset_full['l10_Last Price'] = dataset_full.groupby('Contract Name')['Last Price'].shift(10)

dataset_full['l1_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(1)
dataset_full['l2_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(2)
dataset_full['l3_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(3)
dataset_full['l4_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(4)
dataset_full['l5_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(5)
dataset_full['l6_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(6)
dataset_full['l7_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(7)
dataset_full['l8_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(8)
dataset_full['l9_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(9)
dataset_full['l10_Stock_price'] = dataset_full.groupby('Contract Name')['Stock_price'].shift(10)

dataset_full['l1_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(1)
dataset_full['l2_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(2)
dataset_full['l3_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(3)
dataset_full['l4_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(4)
dataset_full['l5_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(5)
dataset_full['l6_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(6)
dataset_full['l7_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(7)
dataset_full['l8_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(8)
dataset_full['l9_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(9)
dataset_full['l10_Stock_volume'] = dataset_full.groupby('Contract Name')['Stock_volume'].shift(10)

dataset_full.isnull().sum()

37223, 37223*2, 37223*3

dataset_full['l1_Last Price'] = dataset_full['l1_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l2_Last Price'] = dataset_full['l2_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l3_Last Price'] = dataset_full['l3_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l4_Last Price'] = dataset_full['l4_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l5_Last Price'] = dataset_full['l5_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l6_Last Price'] = dataset_full['l6_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l7_Last Price'] = dataset_full['l7_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l8_Last Price'] = dataset_full['l8_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l9_Last Price'] = dataset_full['l9_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))
dataset_full['l10_Last Price'] = dataset_full['l10_Last Price'].fillna(dataset_full.groupby('Contract Name')['Last Price'].transform('median'))


dataset_full['l1_Stock_price'] = dataset_full['l1_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l2_Stock_price'] = dataset_full['l2_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l3_Stock_price'] = dataset_full['l3_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l4_Stock_price'] = dataset_full['l4_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l5_Stock_price'] = dataset_full['l5_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l6_Stock_price'] = dataset_full['l6_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l7_Stock_price'] = dataset_full['l7_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l8_Stock_price'] = dataset_full['l8_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l9_Stock_price'] = dataset_full['l9_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))
dataset_full['l10_Stock_price'] = dataset_full['l10_Stock_price'].fillna(dataset_full.groupby('Contract Name')['Stock_price'].transform('median'))


dataset_full['l1_Stock_volume'] = dataset_full['l1_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l2_Stock_volume'] = dataset_full['l2_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l3_Stock_volume'] = dataset_full['l3_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l4_Stock_volume'] = dataset_full['l4_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l5_Stock_volume'] = dataset_full['l5_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l6_Stock_volume'] = dataset_full['l6_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l7_Stock_volume'] = dataset_full['l7_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l8_Stock_volume'] = dataset_full['l8_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l9_Stock_volume'] = dataset_full['l9_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))
dataset_full['l10_Stock_volume'] = dataset_full['l10_Stock_volume'].fillna(dataset_full.groupby('Contract Name')['Stock_volume'].transform('median'))

dataset_full.isnull().sum()

dataset_full.to_csv('6_More_features_06_05.csv', index=None)

# Complete
